+++ packages/localize-tools/src/tests/transform.unit.test.ts
@@ -13,7 +13,7 @@ import {compileTsFragment, CompilerHostCache} from './compile-ts-fragment.js';
 
 const cache = new CompilerHostCache();
 const IMPORT_MSG = `import { msg, str } from "@lit/localize";\n`;
-const IMPORT_LIT_HTML = `import { html } from "lit-html";\n`;
+const IMPORT_LIT = `import { html } from "lit";\n`;
 
 /**
  * Compile the given fragment of TypeScript source code using the lit-localize
@@ -40,8 +40,8 @@ function checkTransform(
       // should be un-used after litLocalizeTransformation.
     }
     if (inputTs.includes('html')) {
-      inputTs = IMPORT_LIT_HTML + inputTs;
-      expectedJs = IMPORT_LIT_HTML + expectedJs;
+      inputTs = IMPORT_LIT + inputTs;
+      expectedJs = IMPORT_LIT + expectedJs;
     }
   }
   const options = ts.getDefaultCompilerOptions();
@@ -51,6 +51,7 @@ function checkTransform(
   // Don't automatically load typings from nodes_modules/@types, we're not using
   // them here, so it's a waste of time.
   options.typeRoots = [];
+  options.experimentalDecorators = true;
   const result = compileTsFragment(inputTs, options, cache, (program) => ({
     before: [
       litLocalizeTransform(
@@ -444,18 +445,46 @@ test('different variable cast to "lit-localize-status" unchanged', (t) => {
   );
 });
 
-test('Localized(LitElement) -> LitElement', (t) => {
+test('updateWhenLocaleChanges -> undefined', (t) => {
   checkTransform(
     t,
-    `import {LitElement, html} from 'lit-element';
-     import {Localized} from '@lit/localize/localized-element.js';
-     import {msg} from '@lit/localize';
-     class MyElement extends Localized(LitElement) {
+    `import {LitElement, html} from 'lit';
+     import {msg, updateWhenLocaleChanges} from '@lit/localize';
+     class MyElement extends LitElement {
+       constructor() {
+         super();
+         updateWhenLocaleChanges(this);
+       }
        render() {
          return html\`<b>\${msg('Hello World!', {id: 'greeting'})}</b>\`;
        }
      }`,
-    `import {LitElement, html} from 'lit-element';
+    `import {LitElement, html} from 'lit';
+     class MyElement extends LitElement {
+       constructor() {
+         super();
+         undefined;
+       }
+       render() {
+         return html\`<b>Hello World!</b>\`;
+       }
+     }`,
+    {autoImport: false}
+  );
+});
+
+test('@localized removed', (t) => {
+  checkTransform(
+    t,
+    `import {LitElement, html} from 'lit';
+     import {msg, localized} from '@lit/localize';
+     @localized()
+     class MyElement extends LitElement {
+       render() {
+         return html\`<b>\${msg('Hello World!', {id: 'greeting'})}</b>\`;
+       }
+     }`,
+    `import {LitElement, html} from 'lit';
      class MyElement extends LitElement {
        render() {
          return html\`<b>Hello World!</b>\`;