+++ packages/localize-tools/testdata/build-transform-xliff/input/foo.ts
@@ -4,14 +4,15 @@
  * SPDX-License-Identifier: BSD-3-Clause
  */
 
-import {LitElement, html} from 'lit-element';
+import {LitElement, html} from 'lit';
 import {
   msg,
   str,
   configureTransformLocalization,
+  updateWhenLocaleChanges,
+  localized,
   LOCALE_STATUS_EVENT,
 } from '@lit/localize';
-import {Localized} from '@lit/localize/localized-element.js';
 
 const {getLocale} = configureTransformLocalization({sourceLocale: 'en'});
 console.log(`Locale is ${getLocale()}`);
@@ -57,7 +58,20 @@ msg(html`Hello <b><!-- comment -->World</b>!`);
 // Custom ID
 msg('Hello World', {id: 'myId'});
 
-export class MyElement extends Localized(LitElement) {
+// updateWhenLocaleChanges -> undefined
+export class MyElement1 extends LitElement {
+  constructor() {
+    super();
+    updateWhenLocaleChanges(this);
+  }
+  render() {
+    return html`<p>${msg(html`Hello <b>World</b>!`)} (${getLocale()})</p>`;
+  }
+}
+
+// @localized -> removed
+@localized()
+export class MyElement2 extends LitElement {
   render() {
     return html`<p>${msg(html`Hello <b>World</b>!`)} (${getLocale()})</p>`;
   }