+++ packages/localize/examples/runtime/src/x-greeter.ts
@@ -4,11 +4,14 @@
  * SPDX-License-Identifier: BSD-3-Clause
  */
 
-import {LitElement, html} from 'lit-element';
-import {msg} from '@lit/localize';
-import {Localized} from '@lit/localize/localized-element.js';
+import {LitElement, html} from 'lit';
+import {msg, updateWhenLocaleChanges} from '@lit/localize';
 
-export class XGreeter extends Localized(LitElement) {
+export class XGreeter extends LitElement {
+  constructor() {
+    super();
+    updateWhenLocaleChanges(this);
+  }
   render() {
     return html`<p>${msg(html`Hello <b>World</b>!`)}</p>`;
   }