+++ packages/localize/examples/transform/src/x-greeter.ts
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: BSD-3-Clause
  */
 
-import {LitElement, html} from 'lit-element';
+import {LitElement, html} from 'lit';
 import {msg} from '@lit/localize';
 
 export class XGreeter extends LitElement {